weekly update v2.0.0_golden gb6bee2ae-977 2024-03-04
b6bee2ae - [Feat]cvitdl aud detect 1. Add key word detect func 2. Refactor cvitdl aud_detect
15600954 - [cv180x][vi] use CVI_VI_StartPipe ctrl the pipeline(postraw).
9088beb9 - [feat](vi): custom dmesg by linux 4.19
8d288a82 - [sensor] gc1054 driver
c5c63756 - [feat](ispd2): add distortion calibration support
92baca3e - [Feat]cvitdl aud detect 1.Support aud detection
6406e360 - [feat](spifc):add GM25Q64A to support list
b65232ca - [sensor] support ov02b10
f79ecbb6 - [mars_alios][vi] redefine dynamic dpcm
e9e2cb1e - [sensor]Support sensors running on alios 1.Support sc230ai & sc230ai_slave 2.Support sc530ai linear & wdr 3.Support sp2509 4.Modify turnkey_qfn.yaml
2b73eb3f - [turnkey]Vpss crop config 1.Support customized vpss cropping configuration
ff0a3cc8 - [feat]: support sensor BYD_Bf2257CS
fe7cfa8f - [usb] sync master code
93b12473 - [std]adjust cvi_osdrv directory structure
9f2121ac - [cvitdl]Porting cvitdl module 1.Porting cvi_tdl 2.Modify object_detect demo
4aded907 - [feat](gpio):add gpio debounce
fc4f1fd3 - [quick_startup] sync master code
9c723d78 - [sync code] sync yoc7.8 code: 1.improve device driver 2.add csi2d lib 3.support mmu 4.support s/m mode switch 5.improve security subsystem 6.improve audio subsystem 7.improve video vision subsystem 8.improve graphic subsystem 9.support spi nand flash 10.add console uart rvm register
6d38a2b9 - [mars_alios] dump_raw control dpcm dynamic.
48eb1686 - [feat](cvi_test/vpss_test):demo for ldc one frame added
7c5df6a0 - [Sensor] Support SC2336P & SC2336P_1L
ce576181 - [sensor]Sync sc2336_1L from linux
cea0147d - [Feat]modify ldc bin load way.
55bde7fb - [std]Standardize YAML file names to support their features.
184c0888 - [venc]Modify BitRate's type from u16 to u32
3db27822 - [sensor] Sync sc301iot from linux
b6bb5ea2 - [Feat][CV181x][GDC]mesh dump load function

+++ components/cx_lite_record/include/cx/record/snapshot.h
@@ -26,7 +26,7 @@ public:
     
 private:
 
-    enum VidTaskStat : uint8_t {
+    enum VidTaskStat : int {
         VID_TASK_IDLE = 0,
         VID_TASK_RUNNING,
     };
